splunk/requirements.txt
# MCP Server
mcp
fastmcp

# HTTP requests
requests

# Environment variables
python-dotenv
